[REFACTOR] refactor part image
